pyproject.toml
[project]
name = "discord-rag-bot"
version = "0.1.0"
description = "Discord AI chatbot using Retrieval-Augmented Generation (RAG)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord>=2.3.2",
    "openai>=1.63.0",
    "pydantic>=2.10.6",
    "pymilvus>=2.5.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["E501", "SIM", "I"]


